+++ src/s21_sprintf.h
@@ -52,11 +52,11 @@ void s21_fractional_and_integer_part_of_a_number(long long int *num_int,
                                                  bool flag_zero_negative,
                                                  bool flag_minus_num_g);
 int s21_rounding_and_precision_number(
-    double *num, int *multiply, Prototype *prot, double *save_number_for_g,
-    int *e, int *check_g, int *precison, bool flag_zero_plus,
-    bool flag_zero_negative, bool flag_g, int *save_precision_for_rounding_g,
-    bool this_is_used, int *have_precision_g, int *save_precision_g,
-    int *save_degree);
+    double *num, unsigned long int *multiply, Prototype *prot,
+    double *save_number_for_g, int *e, int *check_g, int *precison,
+    bool flag_zero_plus, bool flag_zero_negative, bool flag_g,
+    int *save_precision_for_rounding_g, bool this_is_used,
+    int *have_precision_g, int *save_precision_g, int *save_degree);
 void s21_writing_int_number_with_point(
     long long int *num_int, int *symbol, char *str_int, bool flag_zero_negative,
     int *counter_symbols_str, int *dont_write_number_with_point,
@@ -68,7 +68,8 @@ int s21_concat_fractional_number_with_degree(int num_i, int counter_symbols_str,
                                              Prototype *prot, char symbol_e,
                                              int e, char *str_degree,
                                              int *dont_write_number_with_point);
-void s21_check_fractional_number_for_zeros(int *multiply, char *str_double,
+void s21_check_fractional_number_for_zeros(unsigned long int *multiply,
+                                           char *str_double,
                                            bool flag_minus_num, bool flag_g,
                                            bool this_is_used);
